# docs/swagger.yaml
basePath: /
definitions:
  api.BaseModelResponse:
    properties:
      id:
        type: integer
      is_archived:
        type: boolean
      name:
        type: string
    type: object
  api.BaseModelsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.BaseModelResponse'
        type: array
    type: object
...
  api.ExtendedOrderResponse:
    properties:
      change_amount:
        type: number
      description:
        type: string
      orderInfo:
        $ref: '#/definitions/api.OrderResponse'
      products:
        items:
          $ref: '#/definitions/api.OrderProduct'
        type: array
      received_amount:
        type: number
      statuses:
        items:
          $ref: '#/definitions/api.OrderStatus'
        type: array
    type: object
  api.ExtendedProductResponse:
    properties:
      brand:
        type: string
      brand_id:
        type: integer
      description:
        type: string
      product_type:
        type: string
      product_type_id:
        type: integer
      productInfo:
        $ref: '#/definitions/api.ProductResponse'
      texture:
        type: string
      texture_id:
        type: integer
      volume:
        type: integer
      volume_type:
        type: string
      volume_type_id:
        type: integer
    type: object
  api.OrderClient:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone:
        type: string
    type: object
  api.OrderProduct:
    properties:
      article:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
  api.OrderResponse:
    properties:
      client:
        $ref: '#/definitions/api.OrderClient'
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      total_amount:
        type: number
    type: object
  api.OrderStatus:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  api.OrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/api.OrderResponse'
        type: array
    type: object
  api.ProductResponse:
    properties:
      article:
        type: string
      image_url:
        type: string
      is_archived:
        type: boolean
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
  api.ProductsResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/api.ProductResponse'
        type: array
    type: object
...
  github_com_phenirain_sso_internal_dto_response.Response-api_BaseModelResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.BaseModelResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-api_BaseModelsResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.BaseModelsResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
...
  github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedOrderResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.ExtendedOrderResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedProductResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.ExtendedProductResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-api_OrdersResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.OrdersResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-api_ProductsResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.ProductsResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
...
  github_com_phenirain_sso_internal_dto_response.Response-order_OrderClientsResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/order.OrderClientsResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-order_OrderProductsResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/order.OrderProductsResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-order_OrderStatusesResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/order.OrderStatusesResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
...
  github_com_phenirain_sso_internal_dto_response.Response-string:
    properties:
      data:
        description: Данные ответа
        type: string
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
...
  order.OrderClientsResponse:
    properties:
      clients:
        items:
          $ref: '#/definitions/api.OrderClient'
        type: array
    type: object
  order.OrderProductsResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/api.OrderProduct'
        type: array
    type: object
  order.OrderRequest:
    properties:
      client_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      products:
        items:
          $ref: '#/definitions/api.OrderProduct'
        type: array
      received_amount:
        type: number
      status_id:
        type: integer
    type: object
  order.OrderStatusesResponse:
    properties:
      statuses:
        items:
          $ref: '#/definitions/api.OrderStatus'
        type: array
    type: object
  product.BaseModelRequest:
    properties:
      base_model:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: SSO service API.
  title: SSO API
  version: "1.0"
paths:
...
  /admin/order:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/order.OrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedOrderResponse'
      security:
      - BearerAuth: []
      summary: Create or update order
      tags:
      - admin-order
  /admin/order/{id}:
    delete:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-string'
      security:
      - BearerAuth: []
      summary: Delete order
      tags:
      - admin-order
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedOrderResponse'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - admin-order
  /admin/order/clients:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-order_OrderClientsResponse'
      security:
      - BearerAuth: []
      summary: Get order clients
      tags:
      - admin-order
  /admin/order/products:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-order_OrderProductsResponse'
      security:
      - BearerAuth: []
      summary: Get order products
      tags:
      - admin-order
  /admin/order/statuses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-order_OrderStatusesResponse'
      security:
      - BearerAuth: []
      summary: Get order statuses
      tags:
      - admin-order
  /admin/orders/status/{statusId}:
    get:
      parameters:
      - description: Status ID
        in: path
        name: statusId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_OrdersResponse'
      security:
      - BearerAuth: []
      summary: Get orders
      tags:
      - admin-order
  /admin/product:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_ProductsResponse'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - admin-product
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Product article
        in: formData
        name: article
        type: string
      - description: Old article
        in: formData
        name: article_old
        type: string
      - description: Product name
        in: formData
        name: name
        type: string
      - description: Product description
        in: formData
        name: description
        type: string
      - description: Product price
        in: formData
        name: price
        type: number
      - description: Product quantity
        in: formData
        name: quantity
        type: integer
      - description: Brand ID
        in: formData
        name: brand_id
        type: integer
      - description: Product type ID
        in: formData
        name: product_type_id
        type: integer
      - description: Texture ID
        in: formData
        name: texture_id
        type: integer
      - description: Volume
        in: formData
        name: volume
        type: integer
      - description: Volume ID
        in: formData
        name: volume_id
        type: integer
      - description: Is archived
        in: formData
        name: is_archived
        type: boolean
      - description: Product image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedProductResponse'
      security:
      - BearerAuth: []
      summary: Create or update product
      tags:
      - admin-product
  /admin/product/{article}:
    delete:
      parameters:
      - description: Product article
        in: path
        name: article
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-string'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - admin-product
    get:
      parameters:
      - description: Product article
        in: path
        name: article
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedProductResponse'
      security:
      - BearerAuth: []
      summary: Get product by article
      tags:
      - admin-product
  /admin/product/base-model:
    post:
      consumes:
      - application/json
      parameters:
      - description: Base model request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.BaseModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_BaseModelResponse'
      security:
      - BearerAuth: []
      summary: Create or update base model
      tags:
      - admin-product
  /admin/product/base-model/{baseModelName}/{id}:
    delete:
      parameters:
      - description: Base model name
        in: path
        name: baseModelName
        required: true
        type: string
      - description: Base model ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-string'
      security:
      - BearerAuth: []
      summary: Delete base model
      tags:
      - admin-product
  /admin/product/base-models/{baseModelName}:
    get:
      parameters:
      - description: Base model name
        in: path
        name: baseModelName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_BaseModelsResponse'
      security:
      - BearerAuth: []
      summary: Get all base models
      tags:
      - admin-product
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"